refactor: Folder structure for artist similarity graph
